refactor: simplify schemas
